taylor/pyproject.toml
[project]
name = "taylor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "faiss-cpu>=1.10.0",
    "huggingface-hub==0.14.1",
    "instructorembedding>=1.0.1",
    "langchain>=0.3.18",
    "langchain-chroma>=0.2.2",
    "langchain-community>=0.3.17",
    "langchain-core>=0.3.35",
    "langchain-google-vertexai>=2.0.13",
    "langchain-openai>=0.3.5",
    "langchain-text-splitters>=0.3.6",
    "numpy>=1.26.4",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.0.0",
    "sentence-transformers==2.2.2",
    "setuptools>=75.8.0",
    "streamlit>=1.18.1",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "torchvision>=0.21.0",
    "watchdog>=6.0.0",
]
